+++ .circleci/config.yml
@@ -4,7 +4,7 @@ jobs:
   test:
     docker:
       # Available from https://hub.docker.com/r/circleci/golang/
-      - image: circleci/golang:1.12.5
+      - image: circleci/golang:1.16
     working_directory: /go/src/github.com/bwplotka/mimic
     environment:
       GO111MODULE: 'on'
