+++ .gitignore
@@ -16,3 +16,4 @@
 gen
 .envrc
 vendor
+
