+++ lib/abstr/kubernetes/go.mod
@@ -1,0 +1,9 @@
+module github.com/bwplotka/mimic/lib/abstr/kubernetes
+
+go 1.16
+
+require (
+	github.com/bwplotka/mimic v0.0.0-20190730202618-06ab9976e8ef
+	k8s.io/api v0.20.5
+	k8s.io/apimachinery v0.20.5
+)
