+++ lib/schemas/prometheus/discovery/azure/azure.go
