+++ lib/schemas/prometheus/discovery/consul/consul.go
