+++ lib/schemas/prometheus/discovery/dns/dns.go
