+++ lib/schemas/prometheus/discovery/ec2/ec2.go
