+++ lib/schemas/prometheus/discovery/gce/gce.go
