+++ lib/schemas/prometheus/discovery/kubernetes/kubernetes.go
