+++ lib/schemas/prometheus/discovery/openstack/openstack.go
