+++ lib/schemas/prometheus/discovery/targetgroup/targetgroup.go
