+++ lib/schemas/prometheus/discovery/triton/triton.go
