+++ lib/schemas/prometheus/discovery/zookeeper/zookeeper.go
