+++ lib/schemas/prometheus/go.mod
@@ -1,0 +1,5 @@
+module github.com/bwplotka/mimic/lib/schemas/prometheus
+
+go 1.16
+
+require github.com/prometheus/common v0.20.0
